# infrastructure/swagger/openapi.yml
openapi: 3.0.3
info:
  title: Smart Banking System API
  description: |
    Complete REST API for the TT-BANK Smart Banking System.

    **Architecture:** Microservices — all requests enter through the API Gateway on port 8080.
    The gateway validates the JWT Bearer token before routing to the appropriate backend service.

    **Authentication flow:**
    1. `POST /api/v1/auth/register` — create account
    2. `POST /api/v1/auth/verify-email` — verify OTP sent by email
    3. `POST /api/v1/auth/login` — receive `accessToken` and `refreshToken`
    4. Include `Authorization: Bearer <accessToken>` on all subsequent requests

    **Currency:** Primary currency is **XAF** (Central African CFA franc). Also supports USD, EUR, GBP.

    **Base URL (local dev):** `http://localhost:8080`
    **Base URL (production):** `https://yourdomain.com`
  version: "1.0.0"
  contact:
    name: Smart Banking Dev Team
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: https://yourdomain.com
    description: Production (replace with actual domain)

# =============================================================
# Security scheme — JWT Bearer token (HS256)
# =============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by POST /api/v1/auth/login.
        Include as: Authorization: Bearer <token>

  # =============================================================
  # Reusable schemas — built from actual Java DTOs
  # =============================================================
  schemas:
...
    RegisterRequest:
      type: object
      required: [fullName, email, phoneNumber, password]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
          example: "Jean-Pierre Nkomo"
        email:
          type: string
          format: email
          example: "jeanpierre@example.com"
        phoneNumber:
          type: string
          pattern: '^\\+?[0-9]{8,15}$'
          example: "+237612345678"
        password:
          type: string
          minLength: 8
          format: password
          example: "SecurePass123!"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "jeanpierre@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123!"
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiJ9..."
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiJ9..."
        tokenType:
          type: string
          example: "Bearer"
        userId:
          type: integer
          format: int64
          example: 1
        fullName:
          type: string
          example: "Jean-Pierre Nkomo"
        email:
          type: string
          example: "jeanpierre@example.com"
        role:
          $ref: '#/components/schemas/UserRole'
        expiresIn:
          type: integer
          format: int64
          description: Token validity in milliseconds
          example: 86400000
    VerifyEmailRequest:
      type: object
      required: [email, otp]
      properties:
        email:
          type: string
          format: email
          example: "jeanpierre@example.com"
        otp:
          type: string
          minLength: 6
          maxLength: 6
          example: "482931"
    ResendOtpRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: "jeanpierre@example.com"
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiJ9..."
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: "jeanpierre@example.com"
    ResetPasswordRequest:
      type: object
      required: [email, otp, newPassword]
      properties:
        email:
          type: string
          format: email
          example: "jeanpierre@example.com"
        otp:
          type: string
          minLength: 6
          maxLength: 6
          example: "183947"
        newPassword:
          type: string
          minLength: 8
          format: password
          example: "NewSecurePass456!"
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          example: "OldPass123!"
        newPassword:
          type: string
          minLength: 8
          format: password
          example: "NewPass456!"
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully."
    UserRole:
      type: string
      enum: [USER, MERCHANT, ADMIN]

    # ── Wallet ────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed: email must be a valid email address"
        timestamp:
          type: string
          format: date-time
          example: "2024-06-01T14:30:00"
        path:
          type: string
          example: "/api/v1/auth/register"
  responses:
    Unauthorized:
      description: Missing or invalid JWT Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request body failed validation constraints
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# Default security — applied to every operation unless overridden
security:
  - BearerAuth: []

# =============================================================
# API Paths
# =============================================================
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user account
      description: >
        Creates a new user account. Sends a 6-digit OTP to the provided
        email address. The account is inactive until the OTP is verified
        via POST /api/v1/auth/verify-email.
      security: []   # Public — no token required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created — OTP sent to email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Registration successful. Please check your email for the verification OTP."
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email or phone number already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify email with OTP
      description: Activates the user account using the 6-digit OTP sent to the registered email. OTP expires after 10 minutes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified — account is now active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/resend-verification:
    post:
      tags: [Authentication]
      summary: Resend email verification OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendOtpRequest'
      responses:
        '200':
          description: New OTP sent to email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate and obtain JWT tokens
      description: Returns an access token (24h) and a refresh token (7d). Email must be verified before login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials or email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Issues a new access token using a valid refresh token. Does not require the Bearer header.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Refresh token expired or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Invalidate the current access token
      description: Adds the current access token to the Redis blacklist. All subsequent requests with this token will be rejected.
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get current authenticated user profile
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset OTP
      description: Sends a 6-digit OTP to the user's registered email for password reset.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset OTP sent to email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password using OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/change-password:
    post:
      tags: [Authentication]
      summary: Change password (authenticated)
      description: Requires a valid access token. Changes the password for the currently authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # WALLET SERVICE — /api/v1/wallet/**
  # ============================================================
...