# services/wallet-service/src/main/java/com/example/wallet_service/dto/MessageResponse.java
package com.example.wallet_service.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class MessageResponse {
    private String message;
}